Instances rework, new iSOUPTree models and import cleanup

+++ moa/src/main/java/moa/classifiers/meta/WeightedMajorityAlgorithm.java
@@ -19,19 +19,20 @@
  */
 package moa.classifiers.meta;
 
+import com.github.javacliparser.FlagOption;
+import com.github.javacliparser.FloatOption;
+import com.github.javacliparser.ListOption;
+import com.github.javacliparser.Option;
+import com.yahoo.labs.samoa.instances.Instance;
+
 import moa.classifiers.AbstractClassifier;
 import moa.classifiers.Classifier;
 import moa.core.DoubleVector;
 import moa.core.Measurement;
 import moa.core.ObjectRepository;
-import moa.options.ClassOption;
-import com.github.javacliparser.FlagOption;
-import com.github.javacliparser.FloatOption;
-import com.github.javacliparser.ListOption;
-import com.github.javacliparser.Option;
-import moa.tasks.TaskMonitor;
-import com.yahoo.labs.samoa.instances.Instance;
 import moa.core.Utils;
+import moa.options.ClassOption;
+import moa.tasks.TaskMonitor;
 
 /**
  * Weighted majority algorithm for data streams.
@@ -196,9 +197,9 @@ public class WeightedMajorityAlgorithm extends AbstractClassifier {
         this.ensembleWeights = newEnsembleWeights;
     }
 
-    protected int removePoorestModelBytes() {
+    protected long removePoorestModelBytes() {
         int poorestIndex = Utils.minIndex(this.ensembleWeights);
-        int byteSize = this.ensemble[poorestIndex].measureByteSize();
+        long byteSize = this.ensemble[poorestIndex].measureByteSize();
         discardModel(poorestIndex);
         return byteSize;
     }
